Bitly aware download links
Added functionality for the download links read and parse bitly links. This allows the mod auther to still gain statistics about how many downloads are coming from the update manager via bitly. The downloading works as intended but I am still yet to test if bitly still updates it's statistics as the site is in maintenence mode. I also included a small work around for the '&' symbol as this apearse in the file path in my dropbox.

The code is slightly sloppy, sorry I will clean it up if you want.

nerd-boy